Fixing #17 by setting covariances to zero for nInd=1
We need this when we look at breeding value of a honeybee queen.

It seems that when we have nInd=1, then `arma::cov(X,1)` is casting the 1x2 matrix `cbind(bv, dd)` into a vector and calculating 1x1 variance of `c(bv, dd)`, which causes this code `popVar(cbind(bv[,i],dd[,i]))[1,2]` to break (since there is no 2x2 matrix).